requirements.txt
discord~=2.3.2
python-dotenv==0.19.2
black==24.8.0